chore(dev): mini fixes on admin
