+++ src/bp/admin/ui/public/index.html
@@ -10,8 +10,8 @@
     <base href="/" />
     <script src="%REACT_APP_API_URL%admin/env.js"></script>
     <link rel="manifest" href="%PUBLIC_URL%/manifest.json" />
-    <link rel="icon" href="assets/admin/ui/public/favicon.ico" />
-    <link href="assets/ui-studio/public/external/font-roboto.css" rel="stylesheet" />
+    <link rel="icon" href="%REACT_APP_API_URL%assets/admin/ui/public/favicon.ico" />
+    <link href="%REACT_APP_API_URL%assets/admin/ui/public/fonts/font-roboto.css" rel="stylesheet" />
   </head>
 
   <body>
