+++ src/bp/admin/ui/src/app/style/index.scss
@@ -8,6 +8,12 @@ body {
   margin: 0;
   padding: 0;
   font-family: Roboto !important;
+  font-size: 1rem;
+  font-weight: 400;
+  line-height: 1.5;
+  color: #212529;
+  text-align: left;
+  background-color: #fff;
 }
 
 .container > .navbar {
