+++ src/bp/ui-shared-lite/auth/index.ts
@@ -4,6 +4,7 @@ import { AxiosInstance } from 'axios'
 import { StoredToken, TokenUser, TokenResponse } from 'common/typings'
 import moment from 'moment'
 import ms from 'ms'
+import nanoid from 'nanoid'
 import storage from '../utils/storage'
 
 export const TOKEN_KEY = 'bp/token'
@@ -58,4 +59,24 @@ export const logout = async (getAxiosClient: () => AxiosInstance) => {
   localStorage.removeItem(TOKEN_KEY)
   // need to force reload otherwise the token wont clear properly
   window.location.href = window.location.origin + window['ROOT_PATH']
+}
+
+export const setVisitorId = (userId: string, userIdScope?: string) => {
+  if (typeof userId === 'string' && userId !== 'undefined') {
+    storage.set(userIdScope ? `bp/socket/${userIdScope}/user` : 'bp/socket/user', userId)
+    window.__BP_VISITOR_ID = userId
+  }
+}
+
+export const getUniqueVisitorId = (userIdScope?: string): string => {
+  const key = userIdScope ? `bp/socket/${userIdScope}/user` : 'bp/socket/user'
+
+  let userId = storage.get(key)
+  if (typeof userId !== 'string' || userId === 'undefined') {
+    userId = nanoid(24)
+    storage.set(key, userId)
+  }
+
+  window.__BP_VISITOR_ID = userId
+  return userId
 }
